Use kSecAttrAccount to access the keychain item identifier.
kSecAttrGeneric is supposed to contain a CFDataRef. On iOS, we were getting back the result as a string, but on OS X it is returned as a data ref. Since KeychainItemWrapper stores the exact same value in kSecAttrAccount as kSecAttrGeneric, we may as well use kSecAttrAccount which is supposed to be a string.