IOS-17254: don't crash if API response contains unexpected null. (#612)
When we request an item from the Box API, we request the
collections and collection_memberships fields. The collections
field is an array of collection objects, and the memberships
field is an array of collection membership objects, which have a
collection property which is a collection object.

Recently, a backend bug caused some API responses to contain
null instead of a collection object, either in the collections
array or as the collection property of a collection memberhip.

Most of our JSON parsing code deals with dictionaries, and we have
code designed to handle unexpected nulls, but in these cases,
we weren't handling them, so our app crashed when it received
such a response.

This commit skips trying to parse any null collection objects,
ensuring that if we get a malformed response like that in the
future, we won't crash.