IOS-17522 - remove unnecessary code, which elims a race condition
The `outputStreamHasSpaceAvailable` state was an optimization to
avoid acquring a mutex and calling NSOutputStream.hasSpaceAvailable.
Logic using `outputStreamHasSpaceAvailable` wasn't properly contained
within a mutex, so it was open to race conditions. I don't suspect
this race condition was actually triggering this bug (that will
come in another commit). But this change simplifies the code and
removes a race condition.