Cherry pick of the 2 commits addressing IOS-17522 (#641)
* IOS-17522 - remove unnecessary code, which elims a race condition (#639)

The `outputStreamHasSpaceAvailable` state was an optimization to
avoid acquring a mutex and calling NSOutputStream.hasSpaceAvailable.
Logic using `outputStreamHasSpaceAvailable` wasn't properly contained
within a mutex, so it was open to race conditions. I don't suspect
this race condition was actually triggering this bug (that will
come in another commit). But this change simplifies the code and
removes a race condition.

* IOS-17522 - clear receivedDataBuffer when outputStream is cleared (#640)

IOS-17522 shows writeDataToOutputStream accessing freed memory. This is an
attempt to avoid that. This fix is speculative in that I haven't been able
to repro the same crash. I have confirmed that writeDataToOutputStream
will gracefully fail if it executed with receivedDataBuffer==nil.