Fix a flaky test - test_that_forbidden_403_error_does_not_trigger_logout_notification_for_non_token_requests
This test was flaky due to a subtle (and still not well understood)
interaction between OCMock's 'reject' method and the BOXAPIOperation
finish method. This resulted in random crashes in the finish method.

The solution was to use a BOXAPIOperation `spy` implementation allowing
a test to block until the finish method is *completely* finished. This
technique eliminates the random crashing.

Comments in the code have more details.