IOS-17614 - Fix a memory corruption issue in BOXAPIDataOperation
It's likely that this fix, which appears to fix the a big source of
flakiness in running our test suite, will also fix the top crashing
bug in our product (see IOS-17522).

BOXAPIDataOperation instances *were* being "copied", which resulted
in 2 instances retaining a refereence to the same underlying NSOutputStream.
This lead to a variety of crashes when one Op closed the Stream
while the other Op was attempting to use it.

This fix is somewhat distasteful as copying an BOXAPIDataOperation instance
now modifies the instance being copying (nil'ing out the output stream
reference).

There is much wrong with BOXAuthentifcatedOperation and it's subclasses.
Including fuzzy contracts, lots of duplicated code. None of that is addressed
here to keep the fix small/focused on an attempt to fix the #1 crashed
in our product.