IOS-17000 - Refactor to avoid 3 copies of the same 202 handling (#654)
3 sibling Operation classes each had it's own chunk of code to deal
with 202 responses. This code was basically identical.

A bit of refactoring in these classes along with
the base classes BOXAPIDataOperation and BOXAPIAuthenticatedOperation
now has this retry logic centralized, right next to the similar logic
for the token expired error.

Also noticed a verbose and error prone case stmt. I simplified the pattern
making the code more readable/robust.