fix: Always complete auth actions.
Removing ThreadSafeQueue, as it is only used in this one case, it has
more features than necessary, is not in fact thread safe as written, and
it introduces unnecessary thread hops.

Changing AuthModuleDispatcher, so that it is guaranteed to be retained
until all queued actions are complete (as long as they correctly call
the completion block).

Fixing OAuth2Session to ensure that it calls the auth action completion
block, even in the case it is deallocated. This is unlikely to happen
in practice, but this is unclear from local reasoning.